# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Athena API",
    "version": "1.0.0",
    "description": "REST API for Project Athena - Personal productivity and task management"
  },
  "components": {
    "schemas": {
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Unauthorized"]
          },
          "message": {
            "type": "string",
            "default": "Authentication required"
          }
        },
        "required": ["error"]
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Not found"]
          },
          "message": {
            "type": "string",
            "description": "Resource description"
          }
        },
        "required": ["error"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "description": "Additional error details"
          }
        },
        "required": ["error"]
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Validation error"]
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field that failed validation"
                },
                "message": {
                  "type": "string",
                  "description": "Validation error message"
                }
              },
              "required": ["field", "message"]
            }
          }
        },
        "required": ["error", "details"]
      },
...
      "LinkedIntegrationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LinkedIntegration"
          }
        },
        "required": ["data"],
        "description": "Linked integration"
      },
      "LinkedIntegration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Integration ID"
          },
          "provider": {
            "type": "string",
            "enum": ["linear", "github", "google_calendar", "outlook_calendar", "apple_calendar"],
            "description": "Integration provider",
            "example": "github"
          },
          "externalAccountId": {
            "type": "string",
            "description": "External account identifier"
          },
          "scopes": {
            "type": ["string", "null"],
            "description": "OAuth scopes"
          },
          "metadata": {
            "type": ["object", "null"],
            "additionalProperties": {},
            "description": "Additional metadata"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Connection timestamp",
            "example": "2025-01-09T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2025-01-09T12:00:00Z"
          }
        },
        "required": [
          "id",
          "provider",
          "externalAccountId",
          "scopes",
          "metadata",
          "createdAt",
          "updatedAt"
        ]
      },
      "ConnectIntegrationRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["linear", "github", "google_calendar", "outlook_calendar", "apple_calendar"],
            "description": "Integration provider",
            "example": "github"
          },
          "externalAccountId": {
            "type": "string",
            "description": "External account identifier"
          },
          "accessToken": {
            "type": "string",
            "description": "OAuth access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "OAuth refresh token"
          },
          "tokenExpiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Token expiration (ISO 8601)"
          },
          "scopes": {
            "type": "string",
            "description": "OAuth scopes"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "description": "Additional metadata"
          }
        },
        "required": ["provider", "externalAccountId"]
      },
...
      "LinkedIntegrationsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkedIntegration"
            }
          }
        },
        "required": ["data"],
        "description": "List of linked integrations"
      },
...
      "OAuthAuthorizationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OAuthAuthorization"
          }
        },
        "required": ["data"],
        "description": "OAuth authorization info"
      },
      "OAuthAuthorization": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["linear", "github", "google_calendar", "outlook_calendar", "apple_calendar"],
            "description": "Integration provider",
            "example": "github"
          },
          "authorizationUrl": {
            "type": "string",
            "description": "OAuth authorization URL"
          },
          "configured": {
            "type": "boolean",
            "description": "Whether provider is configured"
          }
        },
        "required": ["provider", "authorizationUrl", "configured"]
      },
      "WebhookResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "processed": {
            "type": "object",
            "properties": {
              "eventType": {
                "type": "string",
                "description": "Webhook event type"
              },
              "entityUpdated": {
                "type": "boolean",
                "description": "Whether entity was updated"
              }
            },
            "required": ["eventType"]
          }
        },
        "required": ["success", "processed"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/integrations/connect": {
      "post": {
        "tags": ["Integrations"],
        "summary": "Connect integration",
        "description": "Connect a new integration.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectIntegrationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Integration connected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedIntegrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "409": {
            "description": "Integration already exists for this provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/integrations/{id}": {
      "delete": {
        "tags": ["Integrations"],
        "summary": "Disconnect integration",
        "description": "Disconnect (remove) an integration.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Integration ID"
            },
            "required": true,
            "description": "Integration ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Integration disconnected successfully"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Integrations"],
        "summary": "Get integration",
        "description": "Get a specific integration by ID.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Integration ID"
            },
            "required": true,
            "description": "Integration ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Integration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedIntegrationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/": {
      "get": {
        "tags": ["Integrations"],
        "summary": "List integrations",
        "description": "List all linked integrations for the authenticated user.",
        "responses": {
          "200": {
            "description": "Integrations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedIntegrationsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/integrations/oauth/{provider}/authorize": {
      "get": {
        "tags": ["Integrations"],
        "summary": "Get OAuth URL",
        "description": "Get OAuth authorization URL for a provider.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["linear", "github", "google_calendar", "outlook_calendar", "apple_calendar"],
              "description": "Integration provider",
              "example": "github"
            },
            "required": true,
            "description": "Integration provider",
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "OAuth redirect URI"
            },
            "required": true,
            "description": "OAuth redirect URI",
            "name": "redirect_uri",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth authorization URL retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid provider or missing redirect_uri",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/webhooks/{provider}": {
      "post": {
        "tags": ["Integrations"],
        "summary": "Handle webhook",
        "description": "Handle incoming webhook from an integration provider.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["linear", "github"],
              "description": "Webhook-enabled provider",
              "example": "github"
            },
            "required": true,
            "description": "Webhook-enabled provider",
            "name": "provider",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid webhook signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}